Committing Work in progress
Tested Branch merge and conlict reconciliation with Unknown-Master.

added more AdminConsole commands.

Added error code notes to a few recvs

Removed some stale code from a few sends

+++ Necromancy.Server/Model/Character.cs
@@ -16,7 +16,7 @@ namespace Necromancy.Server.Model
         public float Z { get; set; }
         public byte viewOffset { get; set; }
         public int battlePose { get; set; }
-        public byte charaPose { get; set; }
+        public int charaPose { get; set; }
 
         public byte movementAnim { get; set; }
         public byte animJumpFall { get; set; }
